CodeClimate de-duplication changes
- Removed overlapping methods fropm older DTO classes
- Refactored JWT class to not use static methods
- Added BasePaymentPayload to Preauth
- Created new base class BasePaymentInfo to remove duplicates from
CreateQrPayload and BasePaymentPayload
- Included coverage report to repo for sonarcloud
